+++ grid2op/Space/SerializableSpace.py
@@ -111,8 +111,8 @@ class SerializableSpace(GridObjects, RandomObject):
 
         self.global_vars = None
 
-        self.shape = self._template_obj.shape()
-        self.dtype = self._template_obj.dtype()
+        self._shape = self._template_obj.shapes()
+        self._dtype = self._template_obj.dtypes()
         self.attr_list_vect = copy.deepcopy(type(self._template_obj).attr_list_vect)
 
         self._to_extract_vect = {}  # key: attr name, value: tuple: (beg_, end_, dtype)
@@ -123,6 +123,14 @@ class SerializableSpace(GridObjects, RandomObject):
             self._to_extract_vect[attr] = (beg_, end_, dtype_)
             beg_ += size
 
+    @property
+    def shape(self):
+        return self._shape
+    
+    @property
+    def dtype(self):
+        return self._dtype
+    
     def _custom_deepcopy_for_copy(self, new_obj):
         RandomObject._custom_deepcopy_for_copy(self, new_obj)
 
@@ -132,10 +140,10 @@ class SerializableSpace(GridObjects, RandomObject):
         new_obj._template_obj = self._template_obj.copy()
         new_obj.n = self.n
         new_obj.global_vars = copy.deepcopy(self.global_vars)
-        new_obj.shape = copy.deepcopy(self.shape)
-        new_obj.dtype = copy.deepcopy(self.dtype)
-        new_obj.attr_list_vect = copy.deepcopy(self.attr_list_vect)
-        new_obj._to_extract_vect = copy.deepcopy(self._to_extract_vect)
+        new_obj._shape = copy.deepcopy(self._shape)
+        new_obj._dtype = copy.deepcopy(self._dtype)
+        new_obj.attr_list_vect = copy.deepcopy(self.attr_list_vect)  # TODO is this necessary, that's class attribute I think
+        new_obj._to_extract_vect = copy.deepcopy(self._to_extract_vect)  # TODO is this necessary, that's class attribute I think
 
     @staticmethod
     def from_dict(dict_):